# Conflicts:
#	src/Game/GameServer.cpp
#	src/Lobby/LobbyServer.cpp
#	src/Lobby/LobbyServer.h
#	src/Networking.h
#	src/Player/PlayerState.cpp
#	src/World/WorldClient.cpp
#	src/World/WorldClient.h
#	src/World/WorldState.cpp
#	src/World/WorldState.h
#	src/main.cpp

+++ src/Player/PlayerState.cpp
@@ -18,7 +18,7 @@ PlayerState::PlayerState(sf::Packet pkt)
 	deserialize(pkt);
 }
 
-void PlayerState::update(float dt)
+void PlayerState::update(int32_t dt)
 {
 	// cooldowns
 	m_shootCooldown.update(dt);
